requirements.txt
pyswisseph
gradio
reportlab
python-dateutil
pytz
